refactor: EventBus as composable
