+++ src/plugins/plot/PlotViewProvider.js
@@ -20,7 +20,6 @@
  * at runtime from the About dialog for additional information.
  *****************************************************************************/
 
-import EventBus from 'utils/eventBus';
 import mount from 'utils/mount';
 
 import Plot from './PlotView.vue';
@@ -77,8 +76,7 @@ export default function PlotViewProvider(openmct) {
               provide: {
                 openmct,
                 domainObject,
-                path: objectPath,
-                eventBus: EventBus
+                path: objectPath
               },
               data() {
                 return {
