--- src/utils/eventBus.js
@@ -1,15 +1,0 @@
-import emitter from 'tiny-emitter/instance';
-import { ref } from 'vue';
-
-// Create a reactive reference to the emitter
-const reactiveEmitter = ref(emitter);
-
-// Expose the emitter's methods
-const EventBus = {
-  $on: (...args) => reactiveEmitter.value.on(...args),
-  $once: (...args) => reactiveEmitter.value.once(...args),
-  $off: (...args) => reactiveEmitter.value.off(...args),
-  $emit: (...args) => reactiveEmitter.value.emit(...args)
-};
-
-export default EventBus;
